Split up dao-factory tests into multiple files
